Committing changes on split group logic before rolling it back.
I've determined that the split group/hole logic is the wrong approach to
the problem of comment groups that are only slightly misaligned from one
another. The strategy adds significant complexity, and I underestimated
the logic required to decide how many groups to chain together with this
mechanism. The *correct* approach would probably warrant its own,
dedicated optimization algorithm, which seems overkill, given the
purpose of this feature.

The approach I plan to implement instead is something along the line of
tabstops for eol comments: a discretization of the alignment columns
using some sort of user-configurable modulus.